Fix values for spatial testing
