+++ community/kernel/src/test/java/org/neo4j/kernel/impl/index/schema/LayoutTestUtil.java
@@ -70,53 +70,7 @@ abstract class LayoutTestUtil<KEY extends NativeSchemaKey, VALUE extends NativeS
     {
     }
 
-    Iterator<IndexEntryUpdate<IndexDescriptor>> randomUpdateGenerator( RandomRule random )
-    {
-        double fractionDuplicates = fractionDuplicates();
-        return new PrefetchingIterator<IndexEntryUpdate<IndexDescriptor>>()
-        {
-            private final Set<Double> uniqueCompareValues = new HashSet<>();
-            private final List<Value> uniqueValues = new ArrayList<>();
-            private long currentEntityId;
-
-            @Override
-            protected IndexEntryUpdate<IndexDescriptor> fetchNextOrNull()
-            {
-                Value value;
-                if ( fractionDuplicates > 0 && !uniqueValues.isEmpty() &&
-                        random.nextFloat() < fractionDuplicates )
-                {
-                    value = existingNonUniqueValue( random );
-                }
-                else
-                {
-                    value = newUniqueValue( random );
-                }
-
-                return add( currentEntityId++, value );
-            }
-
-            private Value newUniqueValue( RandomRule randomRule )
-            {
-                Number value;
-                Double compareValue;
-                do
-                {
-                    value = randomRule.numberPropertyValue();
-                    compareValue = value.doubleValue();
-                }
-                while ( !uniqueCompareValues.add( compareValue ) );
-                Value storableValue = asValue( value );
-                uniqueValues.add( storableValue );
-                return storableValue;
-            }
-
-            private Value existingNonUniqueValue( RandomRule randomRule )
-            {
-                return uniqueValues.get( randomRule.nextInt( uniqueValues.size() ) );
-            }
-        };
-    }
+    abstract Iterator<IndexEntryUpdate<IndexDescriptor>> randomUpdateGenerator( RandomRule random );
 
     Value[] extractValuesFromUpdates( IndexEntryUpdate<IndexDescriptor>[] updates )
     {
